refactor: Move DetailAdapter and DetailViewModel into detail package
